Forgot one fix in ff68e74. [datadriven]

+++ datadriven/src/sgpp/datadriven/application/LearnerScenario.cpp
@@ -156,7 +156,8 @@ solver::SLESolverConfiguration LearnerScenario::getSolverConfigurationFinal() {
 void LearnerScenario::setAdaptivityConfiguration(base::AdpativityConfiguration& adaptConfig) {
   (*this).replaceDictAttr("adaptivity");
   (*this)["adaptivity"].replaceIDAttr("maxLevelType", adaptConfig.maxLevelType_);
-  (*this)["adaptivity"].replaceIDAttr("noPoints", adaptConfig.noPoints_);
+  (*this)["adaptivity"].replaceIDAttr("noPoints",
+                                      static_cast<uint64_t>(adaptConfig.noPoints_));
   (*this)["adaptivity"].replaceIDAttr("numRefinements",
                                       static_cast<uint64_t>(adaptConfig.numRefinements_));
   (*this)["adaptivity"].replaceIDAttr("percent", adaptConfig.percent_);
